Coalesce share modal into export modal
