+++ src/components/header/renderer.tsx
@@ -1,8 +1,8 @@
 import stringify from 'json-stringify-pretty-compact';
 import * as React from 'react';
 import ReactDOM from 'react-dom';
-import { Code, ExternalLink, FileText, GitHub, Grid, HelpCircle, Play, Share2, Trash2, X } from 'react-feather';
-import { Portal, PortalWithState } from 'react-portal';
+import { Code, ExternalLink, FileText, GitHub, Grid, HelpCircle, Play, Trash2, X } from 'react-feather';
+import { PortalWithState } from 'react-portal';
 import { withRouter } from 'react-router-dom';
 import Select from 'react-select';
 import { mapDispatchToProps, mapStateToProps } from '.';
@@ -13,7 +13,6 @@ import ExportModal from './export-modal/index';
 import GistModal from './gist-modal/index';
 import HelpModal from './help-modal/index';
 import './index.css';
-import ShareModal from './share-modal/index';
 
 type Props = ReturnType<typeof mapStateToProps> &
   ReturnType<typeof mapDispatchToProps> & { history: any; showExample: boolean };
@@ -130,13 +129,6 @@ class Header extends React.PureComponent<Props, State> {
       <div className="header-button">
         <ExternalLink className="header-icon" />
         {'Export'}
-      </div>
-    );
-
-    const shareButton = (
-      <div className="header-button">
-        <Share2 className="header-icon" />
-        {'Share'}
       </div>
     );
 
@@ -267,7 +259,6 @@ class Header extends React.PureComponent<Props, State> {
 
     const gist = closePortal => <GistModal closePortal={() => closePortal()} />;
     const exportContent = <ExportModal />;
-    const shareContent = <ShareModal />;
     const helpModal = <HelpModal />;
 
     return (
@@ -295,27 +286,6 @@ class Header extends React.PureComponent<Props, State> {
                       </button>
                     </div>
                     <div className="modal-body">{exportContent}</div>
-                    <div className="modal-footer" />
-                  </div>
-                </div>
-              ),
-            ]}
-          </PortalWithState>
-
-          <PortalWithState closeOnEsc>
-            {({ openPortal, closePortal, onOpen, portal }) => [
-              <span key="0" onClick={openPortal}>
-                {shareButton}
-              </span>,
-              portal(
-                <div className="modal-background" onClick={closePortal}>
-                  <div className="modal modal-top" onClick={e => e.stopPropagation()}>
-                    <div className="modal-header">
-                      <button className="close-button" onClick={closePortal}>
-                        <X />
-                      </button>
-                    </div>
-                    <div className="modal-body modal-hidden">{shareContent}</div>
                     <div className="modal-footer" />
                   </div>
                 </div>
